# docs/swagger.yaml
definitions:
  controllers.Response:
    properties:
      message: {}
    type: object
...
  models.Offers:
    properties:
      comments:
        type: string
      contractor_id:
        type: integer
      created_at:
        type: string
      delivery_time:
        type: string
      id:
        type: integer
      price:
        type: number
      status:
        type: boolean
      tender_id:
        type: integer
      updated_at:
        type: string
    required:
    - contractor_id
    - tender_id
    type: object
  models.OffersRequest:
    properties:
      comments:
        type: string
      contractor_id:
        type: integer
      delivery_time:
        type: string
      price:
        type: number
      status:
        type: boolean
      tender_id:
        type: integer
    required:
    - comments
    - contractor_id
    - delivery_time
    - price
    - tender_id
    type: object
...
info:
  contact:
    email: muhtorhongofurov@gmail.com
    name: Muxtorxon Gofurov
    url: https://github.com/muhtorkhon
  description: Tender Management Golang REST API
  title: Tender Management REST API
  version: "1.0"
paths:
...
  /offers:
    get:
      description: Retrieve a paginated list of all offers.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Offers'
            type: array
        "500":
          description: Failed to fetch offers
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get all offers
      tags:
      - offers
    post:
      consumes:
      - application/json
      description: This endpoint creates a new offer with the provided details.
      parameters:
      - description: Offer Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.OffersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Offers'
        "400":
          description: Failed to parse request body
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Failed to create offer
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Create a new offer
      tags:
      - offers
  /offers/{contractor_id}:
    get:
      description: Retrieve an offer by its ID.
      parameters:
      - description: Contractor ID
        in: path
        name: contractor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Offers'
        "404":
          description: Offer not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get a specific offer
      tags:
      - offers
  /offers/{id}:
    delete:
      description: Mark an offer as deleted by setting the DeletedAt field.
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Offer deleted successfully
          schema:
            type: string
        "404":
          description: Offer not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Soft delete an offer
      tags:
      - offers
    put:
      consumes:
      - application/json
      description: Update the details of an offer by its ID.
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated offer details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.OffersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OffersRequest'
        "400":
          description: Failed to parse request body
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: Offer not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Update an existing offer
      tags:
      - offers
  /offers/{id}/restore:
    patch:
      description: Restore an offer that was previously soft deleted.
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Offer restored successfully
          schema:
            type: string
        "404":
          description: Offer not found
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Restore a soft-deleted offer
      tags:
      - offers
  /offers/filter:
    get:
      consumes:
      - application/json
      description: Retrieves offers with minimum and maximum prices and delivery times,
        along with their counts.
      produces:
      - application/json
      responses:
        "200":
          description: Details of min/max offers and counts
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Error message
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get min, max prices and delivery times with filtered count
      tags:
      - offers
  /offers/sorted:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of offers with pagination, sorted
        by price and delivery time.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of offers per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with offers and total count
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get filtered and sorted offers with pagination
      tags:
      - offers
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"